Refactor et corrections
