+++ src/main/java/cubesolver/IHM/ComposantsUI/BoutonModele.java
@@ -6,8 +6,11 @@ import java.awt.event.ActionListener;
 
 import static cubesolver.IHM.GestionAffichage.couleurBoutons;
 
-public abstract class boutonMouv extends JButton implements ActionListener {
-    public boutonMouv(String mouv, int x, int y, int largeur, int hauteur){
+/**
+ * Classe abstraite définissant le style d'un bouton qui ont tous le même style
+ */
+public abstract class BoutonModele extends JButton implements ActionListener {
+    public BoutonModele(String mouv, int x, int y, int largeur, int hauteur){
         this.setBounds(x, y, largeur, hauteur);
         this.setBackground(couleurBoutons);
         this.setForeground(Color.white);
