+++ src/main/java/cubesolver/IHM/ComposantsUI/BoutonsLancement/ConteneurLancement.java
@@ -1,6 +1,6 @@
-package cubesolver.IHM.ComposantsUI;
+package cubesolver.IHM.ComposantsUI.BoutonsLancement;
 
-import cubesolver.Cube.Cube;
+import cubesolver.IHM.ComposantsUI.BoutonModele;
 import cubesolver.IHM.Etat;
 import cubesolver.IHM.GestionAffichage;
 
@@ -8,18 +8,18 @@ import javax.swing.*;
 import java.awt.*;
 
 
-public class BoutonLancement extends JPanel implements Etat {
-    private static boutonMouv solve;
-    private static boutonMouv mix;
+public class ConteneurLancement extends JPanel implements Etat {
+    private static BoutonModele solve;
+    private static BoutonModele mix;
 
-    public BoutonLancement() {
+    public ConteneurLancement() {
         Color couleurFond = new Color(48, 48, 48);
         this.setLayout(null);
         this.setBounds(1025, 95, 300, 100);
         this.setBackground(couleurFond);
 
-        solve = new boutonResoudre("Résoudre", 0,0,150,100);
-        mix = new boutonMelanger("Mélanger", 150,0,150,100);
+        solve = new BoutonResoudre("Résoudre", 0,0,150,100);
+        mix = new BoutonMelanger("Mélanger", 150,0,150,100);
 
         this.add(solve); this.add(mix);
     }
